+++ lib/preservation/builder.rb
@@ -35,9 +35,9 @@ module Preservation
     # @param directory_name_scheme [Symbol]
     # @return [String]
     def self.build_directory_name(metadata_record, directory_name_scheme)
-      doi = metadata_record['doi']
-      uuid = metadata_record['uuid']
-      title = metadata_record['title'].strip.gsub(' ', '-').gsub('/', '-')
+      doi = metadata_record[:doi]
+      uuid = metadata_record[:uuid]
+      title = metadata_record[:title].strip.gsub(' ', '-').gsub('/', '-')
       time = Time.new
       date = time.strftime("%Y-%m-%d")
       time = time.strftime("%H:%M:%S")
@@ -63,12 +63,12 @@ module Preservation
         when :uuid
           uuid
         when :doi
-          if doi.empty?
+          if doi.nil? || doi.empty?
             return ''
           end
           doi.gsub('/', '-')
         when :doi_short
-          if doi.empty?
+          if doi.nil? || doi.empty?
             return ''
           end
           doi_short_to_remove = 'http://dx.doi.org/'